% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{clean_data}
\alias{clean_data}
\title{Clean and save survey data}
\usage{
clean_data(..., .x = .cluster$cfg$last_data, .wt = NULL)
}
\arguments{
\item{...}{Mutates to clean data.}

\item{.x}{A string; the file path to the data to be used. Defaults to most recent data file.}

\item{.wt}{A function call, such as \code{clean_weight}, to weight data if desired.}
}
\description{
Applies mutates and, if a method is specified, calculates weights for survey
data, then saves the data.
}
\details{
Your function call will likely be quite long, so you should probably save it
in an R script to simplify making minor tweaks. It will quickly become
unwieldy in a console.

If a path to a data file is not specified, the most recent data file will
be used. You can specify another filepath as a string, passed to \code{.x}.

You can specify a function call, such as \code{clean_weight}, as \code{.wt} to weight
data. The function you call should return a function with one argument (the
dataframe of survey data); the created weighting function must return a
dataframe that contains the ID column and any weight columns to add to the
cleaned data.

Other arguments (\code{...}) are passed to \code{dplyr::mutate} for cleaning tasks.
}
