% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_weight.R
\name{clean_weight}
\alias{clean_weight}
\title{Calculate weights from logistic regression and append to data}
\usage{
clean_weight(formula, ..., .import = NA, .filt = T)
}
\arguments{
\item{formula}{The regression formula to pass to \code{glm}.}

\item{...}{Mutates to perform on data before regression; passed to \code{dplyr::mutate}.}

\item{.import}{A string; the file path to a delimited file with cluster-level data to join to the survey data. The file must contain a column titled \code{Cluster} with the cluster identifiers.}

\item{.filt}{Filter data before weighting; defaults to include everyone in the cohort file.}
}
\description{
This function uses logistic regression to weight survey data for analysis.
It should be used as the \code{wt} argument in \code{clean_data}. Returns a dataframe
with inverse probability weights that is appended to your data by
\code{clean_data}. Specifically, it will create a three-column dataframe
with participant IDs, the probability calculated by regression (\code{prob}), and
the inverse probability weight (\code{ipw}).
}
\details{
Technically, \code{clean_weight} returns a function, which is then called by
\code{clean_data} with the survey data as an argument.

You can import data that applies to all clusters by specifying a file as
.import. This file must include a column, \code{Cluster}, with cluster
identifiers consistent with the cohort file. Additional data for cohort
members should be stored in the cohort file, which will be imported and
joined before mutates are applied.

By default, every entry in the cohort file is included for weighting. You
can filter those included for weighting using .filt: for example, you can
use \code{.filt = stringr::str_detect(Status, "Completed")} to only apply
weighting to participants who have completed the survey.
}
