% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_breakdown.R, R/custom_topline.R
\name{custom_topline}
\alias{custom_topline}
\title{Create custom template for breakdown reports}
\usage{
custom_topline()

custom_topline()
}
\description{
Creates an R markdown template in the Scripts folder for you to customize
and use with \code{view_breakdown} rather than its default, simple template.

Creates an R markdown template in the Scripts folder for you to customize
and use with \code{view_topline} rather than its default, simple template.
}
