% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_walklist.R
\name{custom_walklist}
\alias{custom_walklist}
\title{Create custom template for door-to-door walk lists}
\usage{
custom_walklist()
}
\description{
Creates an R markdown template in the Scripts folder for you to customize
and use with \code{make_walklist} rather than its default, simple template.
}
\details{
Your custom template can use additional objects if you pass them as named
arguments to \code{make_walklist}. For example, if you would like to reference
a customizable string representing a time in your template, you can pass it
like \code{make_walklist(template = "Scripts/walklist.Rmd", time = "12:00 AM")}
and refer to it in your template as \code{time}.

If you would like to include additional columns in the walklist tables, you
will need to specify them in \code{make_walklist}, not in a custom template.
}
