% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_cohort.R
\name{export_cohort}
\alias{export_cohort}
\title{Export the cohort file to CSV}
\usage{
export_cohort(
  ...,
  .name = .cluster$cfg$name,
  .status = FALSE,
  .removed = FALSE
)
}
\arguments{
\item{...}{Arguments passed to \code{dplyr::mutate}.}

\item{.name}{Name of the exported file. Defaults to \code{name} from \code{setup}.}

\item{.status}{If TRUE, update status to Enrolled, Re-enroll, or Not enrolled for a new cycle and delete unenrolled/do not contact participants. If FALSE (default), keep current status.}

\item{.removed}{If TRUE, write removed participants to \code{removed.csv}. Defaults to FALSE.}
}
\description{
Generates a CSV file with specified modifications based on the current
cohort file.
}
\details{
\code{export_cohort} does not modify the cohort file itself, only the exported
output. To remove a column from the export using \code{dplyr::mutate} syntax,
set the column to NULL (e.g. \code{export_cohort(Name = NULL)}).

Enabling \code{.status} will prepare the cohort export for a new clusteR
configuration. Previously enrolled participants will remain enrolled,
participants with Unenroll status will be removed, and all others will be
based on their consent status. Participants with \code{NA} consent status are
considered not enrolled and will be included in the export. If you need to
exclude participants from the export that are included by default behavior,
set their status to Unenroll using appropriate mutates.

Enabling \code{.removed} will create a separate file (\code{Cohort/removed.csv}) with
cohort members that are filtered out by \code{.status} when enabled. This is
primarily useful to avoid re-sampling previously sampled households.
}
