% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_data.R
\name{export_data}
\alias{export_data}
\title{Export survey data to CSV}
\usage{
export_data(..., .name = .cluster$cfg$name, .dat = NA, .deid = FALSE)
}
\arguments{
\item{...}{Arguments passed to \code{dplyr::select}. If empty, all columns are kept (except any removed by \code{.deid}).}

\item{.name}{Name of the exported file. Defaults to \code{name} from \code{setup}.}

\item{.dat}{A string; file path to the data archive to use. Defaults to most recent cleaned data if available, or most recent downloaded data otherwise.}

\item{.deid}{If TRUE, removes columns in the Cohort file other than ID. Defaults to FALSE.}
}
\description{
Generates a CSV file from downloaded survey data. By default, uses the most
recent survey data.
}
