% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_csv.R
\name{get_csv}
\alias{get_csv}
\title{get_csv: Get data from a comma-delimited survey data file}
\usage{
get_csv()
}
\description{
Returns a dataframe of survey data from a comma-delimited file specified in
setup.
}
\details{
\code{get_csv} is the simplest data-gathering function. It pulls data from a
.csv input file that you specify during setup. The .csv can generally be
whatever you'd like it to be, but it must contain a few essential fields:
\itemize{
\item ID: The unique ID of the survey participant.
\item Name: The name of the survey participant.
\item Phone: The survey participant's phone number.
\item Email: The survey participant's email address.
\item Consent: The survey participant's consent status.
}

You can format these columns largely however you'd like, and you can even
leave them blank if they are not of use to you, but the columns must exist
and be named correctly.
}
