% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Standardize and save survey data}
\usage{
get_data(args = list())
}
\arguments{
\item{args}{If arguments are needed for your \code{get} function, they can be passed as a list.}
}
\description{
Using the function specified during \code{\link{setup}}, pull and process the latest
survey data, then archive it. Invisibly returns the path to the saved data.
}
\details{
\code{get_data} is the function you should run when you need to download and save
data from your source. It requires proper configuration via \code{setup}, and
obtains data based on parameters set using a \code{setup_get} function.

If you are using a function from a different package to download data for
clusteR, you may wish to attach it using \code{library} before running this
function. If you are using a custom function from an R script to download
data for clusteR, ensure you have set it up properly by including the
script's file path in the clusteR environment as
\code{cluster_cfg$setup_get$run_before}. This should be handled by your custom
\code{setup_get_x} function as described in \code{vignette('setup_get')}.

You should not need to provide any arguments to \code{get_data}. However, if your
get function requires arguments, provide them to \code{get_data} as a named list.
}
