% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_clusters.R
\name{make_clusters}
\alias{make_clusters}
\title{Randomly select census blocks and assign cluster identifiers}
\usage{
make_clusters(n, state = .cluster$cfg$state, county = .cluster$cfg$county)
}
\arguments{
\item{n}{The number of clusters to select.}

\item{state}{The state where clusters will be selected. Defaults to the state in the config file; must be specified if running before \code{setup}.}

\item{county}{The county where clusters will be selected. Defaults to the county in the config file; must be specified if running before \code{setup}.}
}
\description{
Using U.S. Census data, randomly selects Census blocks, weighted by
population, and assign numeric cluster identifiers to them.
}
\details{
Unlike most clusteR functions, \code{make_clusters} can be run before \code{setup}
because you need to identify clusters before you can identify households
to survey.
}
