% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_email.R
\name{make_email}
\alias{make_email}
\title{Generate email list from cohort file}
\usage{
make_email(output, ..., .status = NA)
}
\arguments{
\item{output}{Output format, either "pdf" or "csv".}

\item{...}{Arguments passed to \code{dplyr::filter} to limit the cohort file.}

\item{.status}{If not NA, filters for a status category; see details.}
}
\description{
Produces an email list using the email addresses specified in the cohort file.
Can output a formatted report as a PDF or a CSV for use outside R.
}
\details{
\code{.status} accepts three options:
\itemize{
\item \code{"completed"} includes only residents that have completed the survey
\item \code{"pending"} includes only residents that have \emph{not} completed the survey
\item \code{"enrolled"} includes residents that have enrolled in the cohort, whether
or not they have completed the survey this cycle
}
}
