% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_map.R
\name{make_map}
\alias{make_map}
\title{Generate a customizable map of clusters}
\usage{
make_map(
  x = NA,
  ...,
  layers = c(),
  fill = NULL,
  color = NULL,
  title = NULL,
  subtitle = NULL,
  background = "white"
)
}
\arguments{
\item{x}{Data to dictate cluster fill or border color. Must be a dataframe with column \code{cluster} matching cluster identifiers. Defaults to NA, which will leave clusters empty.}

\item{...}{\code{ggplot2} layers to add to the map. For example, use \code{scale_fill_brewer} to set cluster fill color.}

\item{layers}{A named vector of tigris layers and border colors; see Layers.}

\item{fill}{The column of x (or the Census block shapefile) to fill by, passed to \link[ggplot2:aes]{ggplot2::aes}. \emph{Not} a fill name or scheme function.}

\item{color}{The column of x (or the Census block shapefile) to color borders by, passed to \link[ggplot2:aes]{ggplot2::aes}. \emph{Not} a color name or scheme function.}

\item{title}{The title of your map, if desired.}

\item{subtitle}{The subtitle of your map, if desired.}

\item{background}{The background color for your map. Defaults to white.}
}
\description{
Produces a map of your clusters with layers as requested. For a simpler
function, use \code{view_map}.
}
\section{Layers}{
Additional layers, such as county subdivisions, block groups, and roads, can
be added to maps using their \code{\link[tigris:tigris]{tigris}} functions. Layers
are added using a named vector, like \code{c(function_name = "color")}. Do not use
parentheses with the function name. Examples of layers include:
\itemize{
\item \link[tigris:area_water]{tigris::area_water}
\item \link[tigris:block_groups]{tigris::block_groups}
\item \link[tigris:coastline]{tigris::coastline}
\item \link[tigris:combined_statistical_areas]{tigris::combined_statistical_areas}
\item \link[tigris:congressional_districts]{tigris::congressional_districts}
\item \link[tigris:core_based_statistical_areas]{tigris::core_based_statistical_areas}
\item \link[tigris:county_subdivisions]{tigris::county_subdivisions}
\item \link[tigris:divisions]{tigris::divisions}
\item \link[tigris:landmarks]{tigris::landmarks}
\item \link[tigris:linear_water]{tigris::linear_water}
\item \link[tigris:metro_divisions]{tigris::metro_divisions}
\item \link[tigris:military]{tigris::military}
\item \link[tigris:native_areas]{tigris::native_areas}
\item \link[tigris:new_england]{tigris::new_england}
\item \link[tigris:places]{tigris::places}
\item \link[tigris:primary_roads]{tigris::primary_roads}
\item \link[tigris:primary_secondary_roads]{tigris::primary_secondary_roads}
\item \link[tigris:pumas]{tigris::pumas}
\item \link[tigris:rails]{tigris::rails}
\item \link[tigris:regions]{tigris::regions}
\item \link[tigris:roads]{tigris::roads}
\item \link[tigris:school_districts]{tigris::school_districts}
\item \link[tigris:state_legislative_districts]{tigris::state_legislative_districts}
\item \link[tigris:tracts]{tigris::tracts}
\item \link[tigris:tribal_block_groups]{tigris::tribal_block_groups}
\item \link[tigris:tribal_census_tracts]{tigris::tribal_census_tracts}
\item \link[tigris:tribal_subdivisions_national]{tigris::tribal_subdivisions_national}
\item \link[tigris:urban_areas]{tigris::urban_areas}
\item \link[tigris:voting_districts]{tigris::voting_districts}
\item \link[tigris:zctas]{tigris::zctas}
}

These layers will be automatically trimmed to fit within the boundaries of
your configured county/ies.
}

