% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_walklist.R
\name{make_walklist}
\alias{make_walklist}
\title{Export door-to-door walk lists}
\usage{
make_walklist(
  template = NA,
  cols = c("ID", "Name", "Address", "Consent", "Code", "Notes"),
  ...
)
}
\arguments{
\item{template}{Template file. Defaults to NA.}

\item{cols}{A vector of cohort file column names to be included in the output.}

\item{...}{Objects passed to the environment where the template is built.}
}
\description{
Using groups generated by \code{make_groups} (and manually modified by the user),
generate simple walk lists for in-person surveying. Exports a printable
PDF.
}
\details{
When a cluster's group is specified in the manual column, the manual group
assignment is used. When a cluster's group is not specified, the automatic
group assignment is used. Clusters can be excluded entirely by entering
0 as the manual group.

By default, \code{make_walklist} uses an exceedingly simple template to
generate walklists. You can create and specify your own template R markdown
file in the arguments to generate a custom report, which can be useful for
scripts, reminders, stop times, and other instructional text.
\code{custom_walklist} creates a starter template to customize in your Contacts
folder.
}
