% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_walkmap.R
\name{make_walkmap}
\alias{make_walkmap}
\title{Generate a map of clusters by door-to-door group}
\usage{
make_walkmap(
  save = NA,
  groups = NULL,
  layers = c(),
  title = NULL,
  subtitle = NULL,
  background = "white"
)
}
\arguments{
\item{save}{A string; the file path to save the image.}

\item{groups}{A vector of groups to include.}

\item{layers}{A named vector of tigris layers and border colors; see \code{\link{make_map}}.}

\item{title}{The title of the resulting map, if desired.}

\item{subtitle}{The subtitle of the resulting map, if desired.}

\item{background}{The background color for the map. Defaults to white.}
}
\description{
Using groups generated by \code{make_groups} (and manually modified by the user),
generate simple walk lists for in-person surveying. Exports a printable
PNG image.
}
\details{
When a cluster's group is specified in the manual column, the manual group
assignment is used. When a cluster's group is not specified, the automatic
group assignment is used. Clusters can be excluded entirely by entering
0 as the manual group.

By default, \code{make_walkmap} includes all groups. To include only certain
groups, specify them in a vector as \code{groups}.

By default, \code{make_walkmap} only shows the map in the console. Specifying
a file path in \code{save} will save a PNG image of the map to disk.
}
