% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mult_kmeans.R
\name{mult_kmeans}
\alias{mult_kmeans}
\title{Run multiple iterations of kmeans on cluster geography}
\usage{
mult_kmeans(include, geoids, k, runs, iter.max)
}
\arguments{
\item{include}{A vector of cluster identifiers to include in the map.}

\item{geoids}{A table matching cluster identifiers to Census GEOIDs.}

\item{k}{The number of groups to create.}

\item{runs}{The number of kmeans runs to complete.}

\item{iter.max}{The number of loops each kmeans run should complete.}
}
\description{
Not exported.
}
\keyword{internal}
