% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{partition.object}
\alias{partition.object}
\title{
Partitioning Object
}
\description{
These are objects of class
"partition"
They represent a partitioning of a dataset into clusters.


}
\section{GENERATION}{
This class of objects is returned from \code{pam}, \code{clara} or \code{fanny}.


}
\section{METHODS}{
The \code{"partition"} class has a method for the following generic functions:
\code{plot}, \code{clusplot}.


}
\section{INHERITANCE}{
The following classes inherit from class \code{"partition"} :
\code{"pam"}, \code{"clara"} and \code{"fanny"}.


}
\section{STRUCTURE}{
See \code{pam.object}, \code{clara.object} and \code{fanny.object} for details.
}
\seealso{
\code{\link{pam}}, \code{\link{pam.object}}, \code{\link{clara}}, \code{\link{clara.object}}, \code{\link{fanny}}, \code{\link{fanny.object}}.
}
\keyword{libcluster}
\keyword{cluster}
% Converted by Sd2Rd version 0.3-2.
