% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.agnes}
\alias{print.agnes}
\title{
Use `print()' on an `agnes' object
}
\description{
Prints the ordering of objects, distances between merging clusters,
agglomerative coefficient, and merge structure of an \code{agnes} object.
}
\usage{
print.agnes(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `agnes'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{agnes}}, \code{\link{agnes.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
