% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.clara}
\alias{print.clara}
\title{
Print Method for CLARA Objects
}
\description{
Prints the best sample, medoids, clustering vector and objective function
of \code{clara} object.
}
\usage{
print.clara(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `clara'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{clara}}, \code{\link{clara.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
