% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.diana}
\alias{print.diana}
\title{
Print Method for DIANA Objects
}
\description{
Prints the ordering of objects, diameters of splitted clusters,
and divisive coefficient of a \code{diana} object.
}
\usage{
print.diana(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `diana'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{diana}}, \code{\link{diana.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
