\name{print.dissimilarity}
\title{Print and Summary Methods for Dissimilarity Objects}
\alias{print.dissimilarity}
\alias{summary.dissimilarity}
\description{
  Print or summarize the distances and the attributes of a
  \code{dissimilarity} object.

  These are methods for the functions \code{print()} and \code{summary()} for
  \code{dissimilarity} objects.  See \code{print}, \code{print.default},
  or \code{summary} for the general behavior of these.
}
\usage{
  print.dissimilarity(x, \dots)
summary.dissimilarity(x, \dots)
}
\seealso{
\code{\link{daisy}}, \code{\link{dissimilarity.object}},
\code{\link{print}}, \code{\link{print.default}}. 
}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
 
