% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.fanny}
\alias{print.fanny}
\title{
Print Method for FANNY Objects
}
\description{
Prints the objective function, membership coefficients and clustering vector
of \code{fanny} object.
}
\usage{
print.fanny(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `fanny'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{fanny}}, \code{\link{fanny.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
