% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.pam}
\alias{print.pam}
\title{
Print Method for PAM Objects
}
\description{
Prints the medoids, clustering vector and objective function
of \code{pam} object.
}
\usage{
print.pam(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `pam'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{pam}}, \code{\link{pam.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
