% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.summary.agnes}
\alias{print.summary.agnes}
\title{
Print Method for SUMMARY.AGNES Objects
}
\description{
Prints all components of a \code{agnes} object.
}
\usage{
print.summary.agnes(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `summary.agnes'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{agnes}}, \code{\link{agnes.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
