% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.summary.clara}
\alias{print.summary.clara}
\title{
Use `print()' on a `summary.clara' object
}
\description{
Prints all components of a \code{clara} object.
}
\usage{
print.summary.clara(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `summary.clara'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{clara}}, \code{\link{clara.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{libcluster}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
