% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.summary.diana}
\alias{print.summary.diana}
\title{
Use `print()' on a `summary.diana' object
}
\description{
Prints all components of a \code{diana} object.
}
\usage{
print.summary.diana(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `summary.diana'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{diana}}, \code{\link{diana.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{libcluster}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
