% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.summary.fanny}
\alias{print.summary.fanny}
\title{
Use `print()' on a `summary.fanny' object
}
\description{
Prints all components of a \code{fanny} object.
}
\usage{
print.summary.fanny(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `summary.fanny'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{fanny}}, \code{\link{fanny.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{libcluster}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
