% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.summary.mona}
\alias{print.summary.mona}
\title{
Use `print()' on a `summary.mona' object
}
\description{
Prints all components of a \code{mona} object.
}
\usage{
print.summary.mona(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `summary.mona'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{mona}}, \code{\link{mona.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{libcluster}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
