% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.summary.pam}
\alias{print.summary.pam}
\title{
Use `print()' on a `summary.pam' object
}
\description{
Prints all components of a \code{pam} object.
}
\usage{
print.summary.pam(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `summary.pam'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{pam}}, \code{\link{pam.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{libcluster}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
