% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restore_cohort.R
\name{restore_cohort}
\alias{restore_cohort}
\title{Restore cohort file}
\usage{
restore_cohort(file = NA)
}
\arguments{
\item{file}{The name of a specific archive if different than the most recent file.}
}
\description{
Loads an archived cohort file and, if confirmed, replaces the current file.
}
\details{
\code{restore_cohort} will load the most recent archive by default, which is the
last file (alphabetically) in Cohort/Archive. You can supply the filename
to \code{file} if you want to restore an older version.
}
