% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_get_alc.R
\name{setup_get_alc}
\alias{setup_get_alc}
\title{Set up a connection to an Alchemer survey}
\usage{
setup_get_alc(survey_id, api_token, api_token_secret, codebook, consent_opts)
}
\arguments{
\item{survey_id}{The Survey ID for your Alchemer survey, found after /id/ in the builder URL.}

\item{api_token}{Your API token.}

\item{api_token_secret}{Your API token secret (password).}

\item{codebook}{The file path, as a string, to a tab-delimited text file to key your survey data to standard column names.}

\item{consent_opts}{A named vector, where names are clusteR Consent options and values are survey responses.}
}
\description{
Generates the setup_get item in cluster_cfg for a connection to an Alchemer
survey.
}
\details{
For \code{\link{get_alc}} to properly process survey data, certain information is
required:

\code{survey_id}: The numeric survey identifier, found after /id/ in the URL of
the survey builder page.

\code{api_token}: Your API Key found by clicking your profile icon at the top
right of your Alchemer interface and selecting "Your API Key".

\code{api_token_secret}: Your API Secret Key found by clicking your profile icon
at the top right of your Alchemer interface and selecting "Your API Key".

\code{codebook}: The file path to a \emph{tab}-delimited text file (not
comma-delimited) with two columns: QN and ALC. The QN column should be a
question number or other standard column identifier (such as ID, Name, etc.),
while the ALC column should be the exact text of the question in Alchemer.

Some additional rules apply to \code{codebook}. For multiple-select questions,
you will need to include in your codebook a row for each answer. Each line
should be formatted as follows:
\itemize{
\item The question number (Q##) should be consistent for all responses to a question.
\item The question number should be followed by an underscore and a unique answer number for each response (_##).
\item The question text (ALC column) should be the exact text of the question, followed immediately by a period, then immediately by the exact text of the response.
}

For example:\tabular{ll}{
   QN \tab ALC \cr
   Q01_01 \tab 1. What is your race? (Select all that apply.).Black \cr
   Q01_02 \tab 1. What is your race? (Select all that apply.).Hispanic or Latino \cr
   Q01_03 \tab 1. What is your race? (Select all that apply.).White \cr
}


It is extremely important that all question and answer text appears exactly
as it appears in Alchemer, in full. Shortening the text in ALC will cause
\code{get_alc} to produce nonsensical results or fail.

\code{consent_opts}: A named vector to convert your Alchemer survey's consent
question responses to clusteR's Consent variable. This should be in the
format \code{c(Consent = "Survey response")}, e.g.
\code{c(Yes = "I agree", No = "I don't agree")}. Valid \code{Consent} values are
shown in \code{vignette("setup_cohort")}. For valid \code{Consent} values with spaces,
enclose the name in backticks (`).

Due to a restriction in Alchemer's API, the maximum number of observations
pulled by \code{get_alc} is 9,999. You will need to write a custom function if
this is not adequate for your survey.
}
