\name{summary.agnes}
\alias{summary.agnes}
\alias{print.summary.agnes}
\title{Summary method for `agnes' objects}
\description{Returns (and prints) a summary list for an \code{agnes} object.}
\usage{
\method{summary}{agnes}(x, \dots)
\method{print}{summary.agnes}(x, \dots)
}
\seealso{\code{\link{agnes}}, \code{\link{agnes.object}}.}
\keyword{cluster}
\keyword{print}

