% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{summary.diana}
\alias{summary.diana}
\title{
Summary method for `diana' objects
}
\description{
Returns a summary list for a \code{diana} object.
}
\usage{
summary.diana(x, \dots)
}
\seealso{
\code{\link{diana}}, \code{\link{diana.object}}.
}
\keyword{libcluster}
\keyword{cluster}
% Converted by Sd2Rd version 0.3-2.
