\name{summary.mona}
\alias{summary.mona}
\alias{print.summary.mona}
\title{Summary method for `mona' objects}
\description{Returns (and prints) a summary list for a \code{mona} object.}
\usage{
\method{summary}{mona}(x, \dots)
\method{print}{summary.mona}(x, \dots)
}
\seealso{\code{\link{mona}}, \code{\link{mona.object}}.}
\keyword{cluster}
\keyword{print}
