% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{summary.pam}
\alias{summary.pam}
\title{
Summary Method for `pam' Objects
}
\description{
Returns a summary list for a \code{pam} object.
}
\usage{
summary.pam(x, \dots)
}
\seealso{
\code{\link{pam}}, \code{\link{pam.object}}.
}
\keyword{libcluster}
\keyword{cluster}
% Converted by Sd2Rd version 0.3-2.
