% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{twins.object}
\alias{twins.object}
\title{
Hierarchical Clustering Object
}
\description{
These are objects of class
"twins"
They represent an agglomerative or divisive (polythetic) hierarchical 
clustering of a dataset.


}
\section{GENERATION}{
This class of objects is returned from \code{agnes} or \code{diana}.


}
\section{METHODS}{
The \code{"twins"} class has a method for the following generic function:
\code{pltree}.


}
\section{INHERITANCE}{
The following classes inherit from class \code{"twins"} :
\code{"agnes"} and \code{"diana"}.


}
\section{STRUCTURE}{
See \code{agnes.object} and \code{diana.object} for details.
}
\seealso{
\code{\link{agnes}}, \code{\link{agnes.object}}, \code{\link{diana}}, \code{\link{diana.object}}.
}
\keyword{libcluster}
\keyword{cluster}
% Converted by Sd2Rd version 0.3-2.
