% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_confirm.R
\name{update_confirm}
\alias{update_confirm}
\title{Update cohort file}
\usage{
update_confirm()
}
\arguments{
\item{args}{If arguments are needed for your \code{get} function, they can be passed as a list.}
}
\description{
Gathers data from manual and automatic sources, generates a report with
data conflicts between the current cohort file and the updated data, and
updates the cohort file.
}
\details{
\code{update} is the first of two functions that must be run in series; after
running and verifying the outputs of \code{update}, run \code{update_confirm}.
\code{update} does not actually \emph{perform} the cohort file update, but instead
prepares the data to be updated later while providing a preview of the
changes for your review. \code{update} will set three global values: \code{df_cohort},
\code{df_manual}, and \code{df_source}. These will then be accessed by \code{update_confirm}.

Importantly, if the \code{update} report alerts you to issues in your data
that you subsequently fix, you must run \code{update} again before running
\code{update_confirm} for your changes to take effect.

\code{update} runs \code{get_data} under the hood, meaning that messages will be
displayed in the console as you run it.

You should not need to provide any arguments to \code{update}. However, if your
get function requires arguments, provide them to \code{update} as a named list, as
you would to \code{get_data}.
}
