% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_breakdown.R
\name{view_breakdown}
\alias{view_breakdown}
\title{Calculate results broken down by a variable and generate report}
\usage{
view_breakdown(
  codebook,
  strata,
  data = .cluster$cfg$last_clean,
  make = "pdf",
  template = NA,
  wt = NULL,
  exclude = NA
)
}
\arguments{
\item{codebook}{A string; the file path to a tab-delimited file with question names. See details.}

\item{strata}{A string; the name of the variable by which to break down results.}

\item{data}{The (cleaned) survey data to use. Defaults to the last \code{clean_data} output.}

\item{make}{"pdf" or "csv", the type of output to produce. See details.}

\item{template}{If producing a PDF report from a custom template, a string; the file path to a .Rmd file.}

\item{wt}{The column in \code{data} to use as survey weights. Defaults to equal weight for each response.}

\item{exclude}{Columns (names, as strings) in \code{data} to exclude from weighting. Generally, these are your demographic variables.}
}
\description{
Using archived (by default, cleaned) survey data, automatically combines
columns, calculates frequencies, and applies weights to survey responses
along a specified variable, then outputs a report of data.
}
\details{
Most often, \code{view_breakdown} will be used to calculate survey responses
along a demographic variable. This variable will need to be created during
data cleaning, not within \code{view_breakdown}, and should be named as a string
for \code{strata}.

Selecting PDF generates a clean report with one table per question, one
group of rows per stratum, and one row in each group per answer choice. If
a .Rmd file is specified as \code{template}, a custom template is used to render
the report. The custom template can retrieve data from \code{tables}, which is a
list of tibbles available in the rendering environment. If found in
\code{codebook}, titles and captions are added to the default report and can be
accessed by a custom template as the tibble \code{codebook}; see below for
details.

Selecting CSV generates a single, clean comma-delimited file with one group
of rows per stratum and one row in each group per answer choice. Specifying
\code{template} has no effect other than to produce a warning. Columns found in
\code{codebook} will not be joined to the output by question number.

The file specified as \code{codebook} must be a \strong{tab-delimited} file and include
a \code{QN} column containing the question numbers corresponding to columns in the
cleaned data and a \code{type} column specifying the question type (\code{mc} for
multiple-choice, \code{ms} for multiple-select, \code{no} for numeric, or \code{ft} for
free-text). If you have multiple-select questions on your survey (select all
that apply), your data likely has multiple columns for these questions;
\code{codebook} should only have one entry for the full question. For example,
while your data may have columns \code{Q01_1}, \code{Q01_2}, and \code{Q01_3} for three
answer choices, your \code{codebook} file should only have a \code{Q01} row. For this
reason, \code{codebook} here is different from in \code{setup} and probably can't be
the same file. For PDF output using the default template, you may also
include a \code{title} column, which will be used as the subheading for each
question, and/or a \code{caption} column, which will be added as paragraph text
between the title and the table.  You may include other columns for a custom
template, but they will have no effect if using the default template.
}
