% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_cohort.R
\name{view_cohort}
\alias{view_cohort}
\title{View cohort file}
\usage{
view_cohort()
}
\description{
Quickly opens the cohort file for viewing. Invisibly returns the cohort file
so you can attach it for manual tasks.
}
