% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_map.R
\name{view_map}
\alias{view_map}
\title{Generate a map of clusters}
\usage{
view_map(title = NULL, subtitle = NULL, background = "white")
}
\arguments{
\item{title}{The title of the resulting map, if desired.}

\item{subtitle}{The subtitle of the resulting map, if desired.}

\item{background}{The background color for the map. Defaults to white.}

\item{fill}{The fill color for the clusters. Defaults to light gray.}
}
\description{
Produces a simple, labeled map of your clusters. For a more complex map, use
\code{make_map}.
}
