% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_progress.R
\name{view_progress}
\alias{view_progress}
\title{Generate a report on survey completion and enrollment by cluster}
\usage{
view_progress(
  breaks = c(1, 30, 50, 80, 90),
  colors = c("white", "tomato3", "tomato2", "tomato", "lightblue", "lightblue4")
)
}
\arguments{
\item{breaks}{A vector of numbers as percentages out of 100.}

\item{colors}{A vector of strings representing colors.}
}
\description{
Generates and opens a brief HTML report. The HTML report is standalone and
can be shared as a file.
}
\details{
Included on the report is a map, which color-codes clusters by their
survey completion percentages. The default color-coding scheme is based on
CDC CASPER, which requires 80\% completion per cluster. You can manually
adjust this scheme by adjusting the vectors of breaks and colors. There
should be one more color specified than the number of breaks.
}
