\name{votes.repub}
\alias{votes.repub}
\title{Votes for Republican Candidate in Presidential Elections}
\usage{data(votes.repub)}
\description{
  A data frame with the percents of votes given to the republican
  candidate in presidential elections from 1856 to 1976.  Rows
  represent the 50 states, and columns the 31 elections.
}
\source{
  S. Peterson (1973):
  \emph{A Statistical History of the American Presidential Elections}.
  New York: Frederick Ungar Publishing Co.
  
  Data from 1964 to 1976 is from R. M. Scammon, 
  \emph{American Votes 12}, Congressional Quarterly.
}
\keyword{datasets}

