% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_alc.R
\name{get_alc}
\alias{get_alc}
\title{get_alc: Get data from an Alchemer survey}
\usage{
get_alc()
}
\description{
Returns a dataframe of survey data from an Alchemer survey specified in
setup.
}
\details{
\code{get_alc} pulls data from \href{https://www.alchemer.com/}{Alchemer}, a survey
platform. It is quite flexible, but some key fields must exist after
conversion using the codebook specified in \code{\link{setup_get_alc}}:
\itemize{
\item ID: The unique ID of the survey participant.
\item Name: The name of the survey participant.
\item Phone: The survey participant's phone number.
\item Email: The survey participant's email address.
\item Consent: The survey participant's consent status.
}

You can format these fields largely however you'd like, and you can even
leave them blank if they are not of use to you, but the fields must exist
and be named correctly.
}
