% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_groups.R
\name{make_groups}
\alias{make_groups}
\title{Generate groups of clusters by proximity for door-to-door surveying}
\usage{
make_groups(
  k,
  filt = n > 0,
  fn = "clusteR:::mult_kmeans",
  geoids = .cluster$cfg$geoids,
  shape_county = .cluster$cfg$shape_county,
  shape_block = .cluster$cfg$shape_block,
  runs = 300,
  iter.max = 50,
  ...
)
}
\arguments{
\item{k}{The number of groups to create.}

\item{filt}{How to filter the summary for making the groups; see details.}

\item{fn}{The function to group clusters, by default \code{mult_kmeans}.}

\item{...}{Arguments to be passed to fn.}
}
\description{
Using a kmeans clustering algorithm and Census block latitude and
longitude, calculates nearby clusters and attempts to produce approximately
even groups. Generates \code{Contacts/Assignments.csv}, where the user can
make adjustments before running \code{make_walklist} or \code{make_walkmap} to export
simple door-to-door materials.
}
\details{
You can filter the clusters by the number of completed, pending, and enrolled
participants in each cluster in the \code{filt} argument. The filter defaults to
the numbers in the following columns:
\itemize{
\item \strong{n}: The total number of cohort members in the cluster.
\item \strong{completed}, \strong{completed_pct}: The number or percent (out of 100) of
participants with any Completed status.
\item \strong{pending}, \strong{pending_pct}: The number or percent (out of 100) of
participants with an Enrolled, Re-enroll, or Not enrolled status.
\item \strong{enrolled}, \strong{enrolled_pct}: The number or percent (out of 100) of
participants with a Completed - enrolled or Enrolled status.
}

\code{make_groups} uses the internal function \code{mult_kmeans} by default. You can
write a custom function instead and pass it as an argument to produce data
if needed. \code{make_groups} requires that function output a dataframe with the
following columns:
\itemize{
\item \strong{cluster}, matching the cohort file cluster identifier
\item \strong{geoid}, matching the Census block GEOID
\item \strong{lat} and \strong{long}, the latitude and longtitude of a point within the cluster
\item \strong{ur}, showing whether the block is urban or rural
\item \strong{geometry}, an \emph{sf} geometry field to draw the block shape
}

Alternatively, you can modify the operation of \code{mult_kmeans} by passing
arguments through \code{make_groups}:
\itemize{
\item \strong{k}, required for \code{make_groups} and defines the number of centers (groups)
to generate
\item \strong{runs}, the number of kmeans runs to perform
\item \strong{iter.max}, the maximum number of iterations each kmeans run will perform
}

\code{make_groups} invisibly returns the same table written to
\code{Contacts/Assignments.csv} so you can manipulate it in R if you choose.
}
