% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{partition.object}
\alias{partition.object}
\title{Partitioning Object}
\description{
  The objects of class \code{"partition"}
  represent a partitioning of a dataset into clusters.
}
\section{GENERATION}{
  These objects are returned from \code{pam}, \code{clara} or \code{fanny}.
}
\section{METHODS}{
  The \code{"partition"} class has a method for the following generic functions:
  \code{plot}, \code{clusplot}.
}
\section{INHERITANCE}{
  The following classes inherit from class \code{"partition"} :
  \code{"pam"}, \code{"clara"} and \code{"fanny"}.

  See \code{\link{pam.object}}, \code{\link{clara.object}} and
  \code{\link{fanny.object}} for details. 
}
\seealso{
  \code{\link{pam}}, \code{\link{clara}}, \code{\link{fanny}}.
}
\keyword{cluster}
% Converted by Sd2Rd version 0.3-2.
