\name{print.clara}
\alias{print.clara}
\title{Print Method for CLARA Objects}
\description{
  Prints the best sample, medoids, clustering vector and objective function
  of \code{clara} object.

  This is a method for the function \code{\link{print}()} for objects
  inheriting from class \code{\link{clara}}.
}
\usage{
\method{print}{clara}(x, \dots)
}
\arguments{
  \item{x}{a clara object.}
  \item{\dots}{potential further arguments (require by generic).}
}
\seealso{
  \code{\link{summary.clara}} producing more output;
  \code{\link{clara}}, \code{\link{clara.object}}, \code{\link{print}},
  \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
