\name{print.diana}
\alias{print.diana}
\title{Print Method for DIANA Objects}
\description{
  Prints the ordering of objects, diameters of splitted clusters,
  and divisive coefficient of a \code{diana} object.

  This is a method for the function \code{\link{print}()} for objects
  inheriting from class \code{\link{diana}}.
}
\usage{
\method{print}{diana}(x, \dots)
}
\arguments{
  \item{x}{a diana object.}
  \item{\dots}{potential further arguments (require by generic).}
}
\seealso{
  \code{\link{diana}}, \code{\link{diana.object}}, \code{\link{print}},
  \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}

