% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.dissimilarity}
\alias{print.dissimilarity}
\title{
Use `print()' on a `dissimilarity' object
}
\description{
Prints the distances and the attributes of a \code{dissimilarity} object.
}
\usage{
print.dissimilarity(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `dissimilarity'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{daisy}}, \code{\link{dissimilarity.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{libcluster}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
