\name{print.dissimilarity}
\title{Print and Summary Methods for Dissimilarity Objects}
\alias{print.dissimilarity}
\alias{summary.dissimilarity}
\description{
  Print or summarize the distances and the attributes of a
  \code{dissimilarity} object.

  These are methods for the functions \code{print()} and \code{summary()} for
  \code{dissimilarity} objects.  See \code{print}, \code{print.default},
  or \code{summary} for the general behavior of these.
}
\usage{
  \method{print}{dissimilarity}(x, \dots)
\method{summary}{dissimilarity}(object, \dots)
}
\arguments{
  \item{x, object}{a dissimilarity object.}
  \item{\dots}{potential further arguments (require by generic).}
}
\seealso{
  \code{\link{daisy}}, \code{\link{dissimilarity.object}},
  \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}


