% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.dissimilarity}
\alias{print.dissimilarity}
\alias{summary.dissimilarity}
\title{Print and Summary Methods for Dissimilarity Objects}
\description{
  Prints or summarizes the distances and the attributes of a
  \code{dissimilarity} object.}
\usage{
  print.dissimilarity(x, \dots)
summary.dissimilarity(x, \dots)

methods for the functions \code{print()} and \code{summary()} for
\code{dissimilarity} objects.  See \code{print}, \code{print.default},
or \code{summary} for the general behavior of these.
}
\seealso{
\code{\link{daisy}}, \code{\link{dissimilarity.object}},
\code{\link{print}}, \code{\link{print.default}}. 
}
\keyword{libcluster}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
 
