\name{print.fanny}
\alias{print.fanny}
\title{Print Method for FANNY Objects}
\description{
  Prints the objective function, membership coefficients and clustering vector
  of \code{fanny} object.

  This is a method for the function \code{\link{print}()} for objects
  inheriting from class \code{\link{fanny}}.
}
\usage{
\method{print}{fanny}(x, \dots)
}
\seealso{
  \code{\link{fanny}}, \code{\link{fanny.object}}, \code{\link{print}},
  \code{\link{print.default}}. 
}
\keyword{cluster}
\keyword{print}
