% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{print.mona}
\alias{print.mona}
\title{
Print Method for MONA Objects
}
\description{
Prints the ordering of objects, separation steps, and used variables
of a \code{mona} object.
}
\usage{
print.mona(x, \dots)


This is a method for the function `print()' for objects
inheriting from class `mona'. See `print' or `print.default' for
the general behavior of this function.
}
\seealso{
\code{\link{mona}}, \code{\link{mona.object}}, \code{\link{print}}, \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}
% Converted by Sd2Rd version 0.3-2.
