\name{print.mona}
\alias{print.mona}
\title{Print Method for MONA Objects}
\description{
  Prints the ordering of objects, separation steps, and used variables
  of a \code{mona} object.

  This is a method for the function \code{\link{print}()} for objects
  inheriting from class \code{\link{mona}}.
}
\usage{
\method{print}{mona}(x, \dots)
}
\arguments{
  \item{x}{a mona object.}
  \item{\dots}{potential further arguments (require by generic).}
}
\seealso{
  \code{\link{mona}}, \code{\link{mona.object}}, \code{\link{print}},
  \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}
