\name{print.pam}
\alias{print.pam}
\title{Print Method for PAM Objects}
\description{
  Prints the medoids, clustering vector and objective function
  of \code{pam} object.

  This is a method for the function \code{\link{print}()} for objects
  inheriting from class \code{\link{pam}}.
}
\usage{
\method{print}{pam}(x, \dots)
}
\seealso{
  \code{\link{pam}}, \code{\link{pam.object}}, \code{\link{print}},
  \code{\link{print.default}}.
}
\keyword{cluster}
\keyword{print}

