% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Create folder structure and generate config file}
\usage{
setup(name, short_name, ...)
}
\arguments{
\item{name}{A string, the full name of your project (used for formatted outputs)}

\item{short_name}{A string, the short name of your project (used for file names and other short outputs)}

\item{setup_get}{The output of a function, such as setup_get_csv or setup_get_alc, to set up your data source connection (see \code{vignette('setup_get')})}

\item{setup_cohort}{The name of a properly-formatted cohort input file or a function to generate that file (see \code{vignette('setup_cohort')})}

\item{state}{The FIPS code for the state of interest}

\item{county}{The FIPS code for the county of interest}
}
\description{
Generates a folder structure and config.rds, which holds persistent
information about the survey environment, in the current working directory.
}
\details{
Setup creates the necessary elements of the survey environment (literally,
the structure and contents of the working directory) for a clusteR managed
survey. You may wish to write your setup function call in an R script and
save it in a subfolder called Scripts (which will be created if you don't
create it yourself) so you can document it and modify it later if needed.

While only name and short_name are required for setup to run successfully,
many of the other parameters are required in practice. You will receive
warnings for parameters you did not specify that are soft-required.

\code{setup_get}: setup should see a list that contains information necessary
for a corresponding \code{get} function to retrieve data from a connection.
As a user, you should call a \code{setup_get} function (like \code{\link{setup_get_csv}}),
which will create this list and pass it to setup for you. For a highly
customized setup, you may wish to \code{source} a script to generate this list
instead.

\code{setup_cohort}: setup should see a string denoting a path where a
properly-formatted cohort input file exists. See \code{vignette('setup_cohort')}
for details, but in short, this cohort input file should be tidy and have the
fields ID, Cluster, Name, Mailing, Physical, City, State, ZIP, Phone, Email,
Consent, and Status, as well as any other fields you would like to persist in
your cohort database. You can call a function here to generate the input file
and return the path as a string. The input file must be a .csv, .xls, .xlsx,
or .rds file with the proper file extension.

\code{state}, \code{county}: This package is designed for a local epidemiologist to
perform a cluster sampling survey; therefore, the state and county (or
county-equivalent) should be specified. setup will obtain the 2020 TIGER/Line
shapefiles for your county for you, but you must specify the
\href{https://en.wikipedia.org/wiki/Federal_Information_Processing_Standard_state_code}{state}
and \href{https://en.wikipedia.org/wiki/List_of_United_States_INCITS_codes_by_county}{county}
FIPS/INCITS codes for your area. These are technically not required if you
choose to avoid all mapping functions, but that may cause problems.
}
