% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_get_csv.R
\name{setup_get_csv}
\alias{setup_get_csv}
\title{setup_get_csv: Set up connection to comma-delimited file for survey data}
\usage{
setup_get_csv(file, ...)
}
\arguments{
\item{file}{A string representing a permanent file path to your input .csv file.}

\item{...}{Arguments passed to readr::read_csv.}
}
\description{
Generates the setup_get item in cluster_cfg for a connection to a .csv file.
}
