% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{summary.agnes}
\alias{summary.agnes}
\title{
Summary method for `agnes' objects
}
\description{
Returns a summary list for an \code{agnes} object.
}
\usage{
summary.agnes(x, \dots)
}
\seealso{
\code{\link{agnes}}, \code{\link{agnes.object}}.
}
\keyword{cluster}
% Converted by Sd2Rd version 0.3-2.
