% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{summary.clara}
\alias{summary.clara}
\title{
Summary method for `clara' objects
}
\description{
Returns a summary list for a \code{clara} object.
}
\usage{
summary.clara(x, \dots)
}
\seealso{
\code{\link{clara}}, \code{\link{clara.object}}.
}
\keyword{libcluster}
\keyword{cluster}
% Converted by Sd2Rd version 0.3-2.
