\name{summary.clara}
\alias{summary.clara}
\alias{print.summary.clara}
\title{Summary method for `clara' objects}
\description{Returns (and prints) a summary list for a \code{clara} object.}
\usage{
\method{summary}{clara}(object, \dots)
\method{print}{summary.clara}(x, \dots)
}
\arguments{
  \item{x, object}{a \code{\link{clara}} object.}
  \item{\dots}{potential further arguments (require by generic).}
}
\seealso{\code{\link{clara}}, \code{\link{clara.object}}.}
\keyword{cluster}
\keyword{print}
