\name{summary.diana}
\alias{summary.diana}
\alias{print.summary.diana}
\title{Summary Method for `diana' Objects}
\description{Returns (and prints) a summary list for a \code{diana} object.}
\usage{
\method{summary}{diana}(object, \dots)
\method{print}{summary.diana}(x, \dots)
}
\arguments{
  \item{x, object}{a \code{\link{diana}} object.}
  \item{\dots}{potential further arguments (require by generic).}
}
\seealso{\code{\link{diana}}, \code{\link{diana.object}}.}
\keyword{cluster}
\keyword{print}

