\name{summary.fanny}
\alias{summary.fanny}
\alias{print.summary.fanny}
\title{Summary method for `fanny' objects}
\description{Returns (and prints) a summary list for a \code{fanny} object.}
\usage{
\method{summary}{fanny}(x, \dots)
\method{print}{summary.fanny}(x, \dots)
}
\seealso{\code{\link{fanny}}, \code{\link{fanny.object}}.}
\keyword{cluster}
\keyword{print}

