% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{summary.fanny}
\alias{summary.fanny}
\title{
Summary Method for `fanny' Objects
}
\description{
Returns a summary list for a \code{fanny} object.
}
\usage{
summary.fanny(x, \dots)
}
\seealso{
\code{\link{fanny}}, \code{\link{fanny.object}}.
}
\keyword{libcluster}
\keyword{cluster}
% Converted by Sd2Rd version 0.3-2.
