% %Z%%M% version %I% created %G%
% %Z%%Q%
\name{summary.mona}
\alias{summary.mona}
\title{
Summary Method for `mona' Objects
}
\description{
Returns a summary list for a \code{mona} object.
}
\usage{
summary.mona(x, \dots)
}
\seealso{
\code{\link{mona}}, \code{\link{mona.object}}.
}
\keyword{libcluster}
\keyword{cluster}
% Converted by Sd2Rd version 0.3-2.
