\name{summary.mona}
\alias{summary.mona}
\alias{print.summary.mona}
\title{Summary Method for `mona' Objects}
\description{Returns (and prints) a summary list for a \code{mona} object.}
\usage{
\method{summary}{mona}(object, \dots)
\method{print}{summary.mona}(x, \dots)
}
\arguments{
  \item{x, object}{a \code{\link{mona}} object.}
  \item{\dots}{potential further arguments (require by generic).}
}
\seealso{\code{\link{mona}}, \code{\link{mona.object}}.}
\keyword{cluster}
\keyword{print}
