\name{summary.pam}
\alias{summary.pam}
\alias{print.summary.pam}
\title{Summary Method for PAM Objects}
\description{Summarize a \code{\link{pam}} object and return an object of class
  \code{summary.pam}.
  There's a \code{\link{print}} method for the latter.
}
\usage{
summary.pam(x, \dots)
\method{print}{summary.pam}(x, \dots)
}
\seealso{\code{\link{pam}}, \code{\link{pam.object}}.
}
\keyword{cluster}

