% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_map.R
\name{view_map}
\alias{view_map}
\title{Generate a map of clusters}
\usage{
view_map(title = NULL, subtitle = NULL, fill = "lightgray")
}
\arguments{
\item{title}{The title of the resulting map, if desired.}

\item{subtitle}{The subtitle of the resulting map, if desired.}

\item{fill}{The fill color for the clusters. Defaults to light gray.}
}
\description{
Produces a simple, labeled map of your clusters. Simpler than \code{make_groups},
which is for cluster grouping.
}
