\name{xclara}
\alias{xclara}
\title{Bivariate Data Set with 3 Clusters}
\description{
  An artificial data set consisting of 3000 points in 3 well-separated
  clusters of size 1000 each.
}
\usage{data(xclara)}
\format{
  A data frame with 3000 observations on 2 numeric variables giving the
  \eqn{x} and \eqn{y} coordinates of the points, respectively.
}
\source{
  Sample data set accompanying the reference below.
}
\references{
  Anja Struyf, Mia Hubert & Peter J. Rousseeuw (1996)
  Clustering in an Object-Oriented Environment.
  \emph{Journal of Statistical Software} \bold{1}.
  \url{http://www.jstatsoft.org/v01/i04}
}
\keyword{datasets}
